feat: added --seeder & --force on db:seed

+++ commands/db/DbSeedCommand.js
@@ -1,4 +1,5 @@
 import Chalk from "@bejibun/logger/facades/Chalk";
+import { ask, defineValue, isNotEmpty } from "@bejibun/utils";
 import ora from "ora";
 import path from "path";
 import Database from "../../facades/Database";
@@ -20,7 +21,10 @@ export default class DbSeedCommand {
      *
      * @var $options Array<Array<any>>
      */
-    $options = [];
+    $options = [
+        ["-f, --force", "Skip command confirmation"],
+        ["-s, --seeder", "Specific file seeder"]
+    ];
     /**
      * The arguments of the console command.
      *
@@ -29,23 +33,36 @@ export default class DbSeedCommand {
     $arguments = [];
     async handle(options, args) {
         const database = Database.knex();
-        const spinner = ora(Chalk.setValue("Seeding...")
-            .info()
-            .show()).start();
-        try {
-            const logs = (await database.seed.run()).flat();
-            spinner.succeed("Seeding finished");
-            if (logs.length > 0)
-                logs.forEach((seeder) => spinner.succeed(path.basename(seeder)));
-            else
-                spinner.succeed("No seeders were run.");
-        }
-        catch (error) {
-            spinner.fail(`Seeding failed : ${error.message}`);
-        }
-        finally {
-            await database.destroy();
-            spinner.stop();
+        const environment = defineValue(Bun.env.APP_ENV, "development");
+        const bypass = isNotEmpty(options.force);
+        const seeder = options.seeder;
+        let confirm = "Y";
+        if (environment === "production" && !bypass)
+            confirm = await ask(Chalk.setValue("Application in production. Are you sure you want to run this command? (Y/N): ")
+                .inline()
+                .error()
+                .show());
+        if (confirm.toUpperCase() === "Y") {
+            const spinner = ora(Chalk.setValue("Seeding...")
+                .info()
+                .show()).start();
+            try {
+                const logs = (await database.seed.run({
+                    specific: defineValue(seeder)
+                })).flat();
+                spinner.succeed("Seeding finished");
+                if (logs.length > 0)
+                    logs.forEach((seeder) => spinner.succeed(path.basename(seeder)));
+                else
+                    spinner.succeed("No seeders were run.");
+            }
+            catch (error) {
+                spinner.fail(`Seeding failed : ${error.message}`);
+            }
+            finally {
+                await database.destroy();
+                spinner.stop();
+            }
         }
     }
 }
